+++ NEWS.md
@@ -1,5 +1,6 @@
 # 0.1-3
 1. Added function `ld_r`.
+2. Added support `s_coefficient` function. Only works with MAF changes file.
 
 # 0.1-2
 1. Discard rows in closest files that have a dot (.) in the gene id column.